Refactoring events to verify in mid-event and generate documents on submit
