+++ src/functionalTests/java/uk/gov/hmcts/reform/finrem/functional/caseorchestration/ValidateHearingDatesTest.java
@@ -2,16 +2,15 @@ package uk.gov.hmcts.reform.finrem.functional.caseorchestration;
 
 import io.restassured.response.Response;
 import net.serenitybdd.junit.runners.SerenityRunner;
-import net.serenitybdd.rest.SerenityRest;
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Value;
 import uk.gov.hmcts.reform.finrem.functional.IntegrationTestBase;
 
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.empty;
+import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.startsWith;
-import static org.junit.Assert.assertEquals;
 
 @RunWith(SerenityRunner.class)
 public class ValidateHearingDatesTest extends IntegrationTestBase {
@@ -24,55 +23,39 @@ public class ValidateHearingDatesTest extends IntegrationTestBase {
 
     @Test
     public void verifyShouldReturnBadRequestWhenCaseDataIsMissingInRequest() {
-        assertEquals(400, utils.getStatusCode(validateHearing, "empty-casedata1.json",consentedDir));
+        Response response = utils.getResponse(validateHearing, "empty-casedata1.json",consentedDir);
 
-        assertThat(utils.getResponse(validateHearing, "empty-casedata1.json",consentedDir).body().asString(),
-            startsWith("Some server side exception occurred. Please check logs for details"));
+        assertThat(response.getStatusCode(), is(400));
+        assertThat(response.body().asString(), startsWith("Some server side exception occurred. Please check logs for details"));
     }
 
     @Test
     public void verifyShouldThrowErrorWhenIssueDateAndHearingDateAreEmpty() {
-
-        assertEquals("Issue Date, fast track decision or hearingDate is empty",
-                getResponse(validateHearing, "pba-validate1.json",consentedDir).jsonPath().get("errors[0]"));
+        assertThat(getResponseAndAssertSuccessStatus(validateHearing, "pba-validate1.json",consentedDir).jsonPath().get("errors[0]"),
+            is("Issue Date, fast track decision or hearingDate is empty"));
     }
 
     @Test
     public void verifyShouldThrowWarningsWhenNotFastTrackDecision() {
-
-        assertEquals("Date of the hearing must be between 12 and 16 weeks.",
-                getResponse(validateHearing, "validate-hearing-without-fastTrackDecision1.json",
-                        contestedDir).jsonPath().get("warnings[0]"));
+        assertThat(getResponseAndAssertSuccessStatus(validateHearing, "validate-hearing-without-fastTrackDecision1.json",
+            contestedDir).jsonPath().get("warnings[0]"), is("Date of the hearing must be between 12 and 16 weeks."));
     }
 
     @Test
     public void verifyshouldThrowWarningsWhenFastTrackDecision() {
-
-        assertEquals("Date of the Fast Track hearing must be between 6 and 10 weeks.",
-                getResponse(validateHearing, "validate-hearing-with-fastTrackDecision1.json",
-                        contestedDir).jsonPath().get("warnings[0]"));
+        assertThat(getResponseAndAssertSuccessStatus(validateHearing, "validate-hearing-with-fastTrackDecision1.json",
+            contestedDir).jsonPath().get("warnings[0]"), is("Date of the Fast Track hearing must be between 6 and 10 weeks."));
     }
 
     @Test
     public void verifyShouldSuccessfullyValidate() {
-
-        Assert.assertNull(
-                getResponse(validateHearing, "validate-hearing-successfully1.json",
-                        contestedDir).jsonPath().get("warnings"));
+        assertThat(getResponseAndAssertSuccessStatus(validateHearing, "validate-hearing-successfully1.json",
+            contestedDir).jsonPath().get("warnings"), is(empty()));
     }
 
-    private Response getResponse(String url,  String jsonFileName, String journeyType) {
-
-        Response response = SerenityRest.given()
-            .relaxedHTTPSValidation()
-            .headers(utils.getHeaders())
-            .body(utils.getJsonFromFile(jsonFileName, journeyType))
-            .when().post(url).andReturn();
-
-        int statusCode = response.getStatusCode();
-        assertEquals(statusCode, 200);
-
+    private Response getResponseAndAssertSuccessStatus(String url, String jsonFileName, String journeyType) {
+        Response response = utils.getResponse(url, jsonFileName, journeyType);
+        assertThat(response.getStatusCode(), is(200));
         return response;
-
     }
 }
