+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/config/ObjectMapperConfigurationTest.java
@@ -7,9 +7,9 @@ import org.mockito.Spy;
 import org.mockito.junit.MockitoJUnitRunner;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
-import static org.junit.Assert.assertThat;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ObjectMapperConfigurationTest {
