+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/NotificationsControllerTest.java
@@ -17,7 +17,6 @@ import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.AssignedToJudgeDocumentService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.BulkPrintService;
-import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.GeneralEmailService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.HelpWithFeesDocumentService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.ManualPaymentDocumentService;
@@ -41,6 +40,7 @@ import static uk.gov.hmcts.reform.finrem.caseorchestration.TestConstants.AUTH_TO
 @RunWith(SpringRunner.class)
 @WebMvcTest(NotificationsController.class)
 public class NotificationsControllerTest {
+
     //URLs
     private static final String HWF_SUCCESSFUL_CALLBACK_URL = "/case-orchestration/notify/hwf-successful";
     private static final String ASSIGN_TO_JUDGE_CALLBACK_URL = "/case-orchestration/notify/assign-to-judge";
@@ -48,7 +48,6 @@ public class NotificationsControllerTest {
     private static final String CONSENT_ORDER_MADE_URL = "/case-orchestration/notify/consent-order-made";
     private static final String ORDER_NOT_APPROVED_URL = "/case-orchestration/notify/order-not-approved";
     private static final String CONSENT_ORDER_AVAILABLE_URL = "/case-orchestration/notify/consent-order-available";
-    private static final String PREPARE_FOR_HEARING_CALLBACK_URL = "/case-orchestration/notify/prepare-for-hearing";
     private static final String PREPARE_FOR_HEARING_ORDER_SENT_CALLBACK_URL = "/case-orchestration/notify/prepare-for-hearing-order-sent";
     private static final String CONTESTED_APPLICATION_ISSUED_CALLBACK_URL = "/case-orchestration/notify/contest-application-issued";
     private static final String CONTEST_ORDER_APPROVED_CALLBACK_URL = "/case-orchestration/notify/contest-order-approved";
@@ -76,29 +75,14 @@ public class NotificationsControllerTest {
     private static final String GENERAL_EMAIL_CONTESTED = "/fixtures/contested/general-email-contested.json";
     private static final String CONTESTED_PAPER_CASE_JSON = "/fixtures/contested/paper-case.json";
 
-    @Autowired
-    private WebApplicationContext applicationContext;
+    @Autowired private WebApplicationContext applicationContext;
 
-    @MockBean
-    private NotificationService notificationService;
-
-    @MockBean
-    private ManualPaymentDocumentService manualPaymentDocumentService;
-
-    @MockBean
-    private FeatureToggleService featureToggleService;
-
-    @MockBean
-    private BulkPrintService bulkPrintService;
-
-    @MockBean
-    private AssignedToJudgeDocumentService assignedToJudgeDocumentService;
-
-    @MockBean
-    private GeneralEmailService generalEmailService;
-
-    @MockBean
-    private HelpWithFeesDocumentService helpWithFeesDocumentService;
+    @MockBean private NotificationService notificationService;
+    @MockBean private ManualPaymentDocumentService manualPaymentDocumentService;
+    @MockBean private BulkPrintService bulkPrintService;
+    @MockBean private AssignedToJudgeDocumentService assignedToJudgeDocumentService;
+    @MockBean private GeneralEmailService generalEmailService;
+    @MockBean private HelpWithFeesDocumentService helpWithFeesDocumentService;
 
     private MockMvc mockMvc;
     private JsonNode requestContent;
@@ -330,30 +314,6 @@ public class NotificationsControllerTest {
         mockMvc.perform(post(HWF_SUCCESSFUL_CALLBACK_URL)
             .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
             .content(requestContent.toString())
-            .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk());
-
-        verifyNoInteractions(notificationService);
-    }
-
-    @Test
-    public void shouldSendPrepareForHearingEmailWhenAgreed() throws Exception {
-        buildCcdRequest(CONTESTED_SOL_SUBSCRIBED_FOR_EMAILS_JSON);
-        mockMvc.perform(post(PREPARE_FOR_HEARING_CALLBACK_URL)
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-            .contentType(MediaType.APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk());
-
-        verify(notificationService, times(1)).sendPrepareForHearingEmail(any());
-    }
-
-    @Test
-    public void shouldNotSendPrepareForHearingEmailWhenNotAgreed() throws Exception {
-        buildCcdRequest(CCD_REQUEST_JSON);
-        mockMvc.perform(post(PREPARE_FOR_HEARING_CALLBACK_URL)
-            .content(requestContent.toString())
-            .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
             .contentType(MediaType.APPLICATION_JSON_VALUE))
             .andExpect(status().isOk());
